fix: fix data type from long to int

+++ src/main/java/br/com/template/service/UserService.java
@@ -5,12 +5,12 @@ import java.util.List;
 import br.com.template.domain.User;
 
 public interface UserService {
-    User getById(long id);
+    User getById(int id);
     
     List<User> getAll();
     
     User create(User user);
     
-    void delete(long id);
+    void delete(int id);
 
 }
